requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-redis
gunicorn
drf-spectacular
whitenoise